General Ver./requirements.txt
gdown==4.7.1
matplotlib==3.7.1
networkx==3.1
NeuroGraph==3.0.0
nni==2.10
numpy==1.24.2
ogb==1.3.6
paddlepaddle_gpu==2.5.2.post117
pandas==2.0.0
scikit_learn==1.2.2
scipy==1.12.0
torch==2.0.0+cu117.with.pypi.cudnn
torch_geometric==2.3.0
torchmetrics==1.2.0
tqdm==4.65.0